+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- Allow `impl_vertex!` to support generic structs.
+
 # Version 0.7.1 (2017-09-28)
 
 - Fixed an HiDPI scaling issue on MacOS in vulkano-win.