+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- Allow `impl_vertex!` to support generic structs.
 - Fixed creating a buffer view not checking the `min_texel_buffer_offset_alignment` limit.
 
 # Version 0.7.1 (2017-09-28)