+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - Allow `impl_vertex!` to support generic structs.
+- Added `GraphicsPipelineBuilder::with_pipeline_layout`.
 - Fixed creating a buffer view not checking the `min_texel_buffer_offset_alignment` limit.
 - Added support for loading the `VK_EXT_debug_marker` extension and adding debug markers to
   `UnsafeCommandBufferBuilder`